+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Referenceable Scriptables
 ## Changelog
 
+1.1.0
+- Referenceables Management editor window added
+- Removing from table deletes empty folders
+- Other code cleanups/stability
+
 1.0.0
 - Basic features implemented and ready!